+++ src/main/java/com/gabriel/muramasa/services/FollowService.java
@@ -4,8 +4,10 @@
  */
 package com.gabriel.muramasa.services;
 
+import com.auth0.jwt.exceptions.TokenExpiredException;
 import com.gabriel.muramasa.handlers.exceptions.DatabaseException;
 import com.gabriel.muramasa.handlers.exceptions.NotFoundException;
+import com.gabriel.muramasa.handlers.exceptions.UnauthorizedException;
 import com.gabriel.muramasa.repositories.AccountRepository;
 import com.gabriel.muramasa.repositories.FollowerRepository;
 import com.gabriel.muramasa.models.Follower;
@@ -28,41 +30,52 @@ public class FollowService {
     @Autowired
     private AccountRepository accRepo;
     
+    private String currentUser;
+    
+    public void setCurrentUser(String currentUser) {
+        this.currentUser = currentUser;
+    }
+    
     public List<Follower> getFollowing(Long userId) {
         Account acc = accRepo.findById(userId).orElseThrow(() -> new NotFoundException("User not found."));
         return acc.getFollowing();
     }
     
     public List<Follower> getFollowers(Long userId) {
+        
         Account acc = accRepo.findById(userId).orElseThrow(() -> new NotFoundException("User not found."));
         return acc.getFollowers();
     }
     
     public Follower follow(Long fromId, Long toId) {
         Account from = accRepo.findById(fromId).orElseThrow(() -> new NotFoundException("From account not found."));
-        Account to = accRepo.findById(toId).orElseThrow(() -> new NotFoundException("To account not found."));
-        
-        Follower follower = new Follower(from, to);
-        
-        try {
-            return repo.save(follower);
-        }catch(ConstraintViolationException e) {
-            throw new DatabaseException(e.getMessage());
-        }
+        if(from.getUsername().equals(this.currentUser)) {
+            Account to = accRepo.findById(toId).orElseThrow(() -> new NotFoundException("To account not found."));
+
+            Follower follower = new Follower(from, to);
+
+            try {
+                return repo.save(follower);
+            }catch(ConstraintViolationException e) {
+                throw new DatabaseException(e.getMessage());
+            }
+        }else throw new UnauthorizedException("Unauthorized operation.");
     }
     
     public Follower unfollow(Long fromId, Long toId) {
         Account from = accRepo.findById(fromId).orElseThrow(() -> new NotFoundException("From account not found."));
-        Account to = accRepo.findById(toId).orElseThrow(() -> new NotFoundException("To account not found."));
-        
-        Follower follower = repo.findBySearchParameter("" + from.hashCode() + to.hashCode()).orElseThrow(() -> new RuntimeException("Some error."));
-        
-        try {
-            repo.delete(follower);
-            return follower;
-        }catch(ConstraintViolationException e) {
-            throw new DatabaseException(e.getMessage());
-        }
+        if(from.getUsername().equals(this.currentUser)) {
+            Account to = accRepo.findById(toId).orElseThrow(() -> new NotFoundException("To account not found."));
+
+            Follower follower = repo.findBySearchParameter("" + from.hashCode() + to.hashCode()).orElseThrow(() -> new RuntimeException("Some error."));
+
+            try {
+                repo.delete(follower);
+                return follower;
+            }catch(ConstraintViolationException e) {
+                throw new DatabaseException(e.getMessage());
+            }
+        }else throw new UnauthorizedException("Unauthorized operation.");
     }
 
 }
