+++ src/main/java/com/gabriel/muramasa/services/PostService.java
@@ -7,6 +7,7 @@ package com.gabriel.muramasa.services;
 import com.gabriel.muramasa.handlers.exceptions.DatabaseException;
 import com.gabriel.muramasa.handlers.exceptions.InvalidFormatException;
 import com.gabriel.muramasa.handlers.exceptions.NotFoundException;
+import com.gabriel.muramasa.handlers.exceptions.UnauthorizedException;
 import com.gabriel.muramasa.repositories.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -40,7 +41,13 @@ public class PostService {
     
     private SimpleDateFormat formatter = new SimpleDateFormat("dd/MM/yyyy HH:mm:ss");
     
+    private String currentUser;
+    
     public PostService() {}
+    
+    public void setCurrentUser(String currentUser) {
+        this.currentUser = currentUser;
+    }
     
     public List<Post> getFollowingPosts(Long userId) {
         Account acc = accService.findById(userId);
@@ -65,38 +72,44 @@ public class PostService {
         String[] words = post.getText().split(" ");
         String[] imgs = post.getAttachedImgs().split(";");
         Account acc = accService.findById(userId);
-        if(words.length > 500) {throw new InvalidFormatException("Post text surpass limit of 500 words.");}
-        if(imgs.length > 4) {throw new InvalidFormatException("Post attached images surpass limit of 4 images.");}
-        post.setCreator(acc);
-        post.setLikes(new ArrayList<Like>());
-        post.setReplies(new ArrayList<Reply>());
-        post.setDate(formatter.format(new Date()));
-        try {
-            Post savedPost = repo.save(post);
-            Log log = new Log(formatter.format(new Date()), "" + acc.getUsername() + " created a post.", acc);
-            logService.addRecentUpdate(log, acc.getRecentUpdates());
-            return savedPost;
-        } catch(ConstraintViolationException e) {
-            throw new DatabaseException(e.getMessage());
-        } catch(ParseException e) {
-            throw new DatabaseException(e.getMessage());
-        }
+        if(acc.getUsername().equals(this.currentUser)) {
+            if(words.length > 500) {throw new InvalidFormatException("Post text surpass limit of 500 words.");}
+            if(imgs.length > 4) {throw new InvalidFormatException("Post attached images surpass limit of 4 images.");}
+            post.setCreator(acc);
+            post.setLikes(new ArrayList<Like>());
+            post.setReplies(new ArrayList<Reply>());
+            post.setDate(formatter.format(new Date()));
+            try {
+                Post savedPost = repo.save(post);
+                Log log = new Log(formatter.format(new Date()), "" + acc.getUsername() + " created a post.", acc);
+                logService.addRecentUpdate(log, acc.getRecentUpdates());
+                return savedPost;
+            } catch(ConstraintViolationException e) {
+                throw new DatabaseException(e.getMessage());
+            } catch(ParseException e) {
+                throw new DatabaseException(e.getMessage());
+            }
+        }else throw new UnauthorizedException("Unauthorized operation.");
     }
     
     public Post update(Post newPost) {
         Post oldPost = repo.findById(newPost.getId()).orElseThrow(() -> new NotFoundException("Post not found."));
-        
-        if(!newPost.getText().equals(oldPost.getText())) {
-            oldPost.setText(newPost.getText());
-        }
-        if(!newPost.getAttachedImgs().equals(oldPost.getAttachedImgs())) {
-            oldPost.setAttachedImgs(newPost.getAttachedImgs());
-        }
-        
-        return repo.save(oldPost);
+        if(oldPost.getCreator().getUsername().equals(this.currentUser)) {
+            if(!newPost.getText().equals(oldPost.getText())) {
+                oldPost.setText(newPost.getText());
+            }
+            if(!newPost.getAttachedImgs().equals(oldPost.getAttachedImgs())) {
+                oldPost.setAttachedImgs(newPost.getAttachedImgs());
+            }
+
+            return repo.save(oldPost);
+        }else throw new UnauthorizedException("Unauthorized operation.");
     }
     
     public void delete(Long postId) {
-        repo.deleteById(postId);
+        Post post = repo.findById(postId).orElseThrow(() -> new NotFoundException("Post not found."));
+        if(post.getCreator().getUsername().equals(this.currentUser)) {
+            repo.deleteById(postId);
+        }else throw new UnauthorizedException("Unauthorized operation.");
     }
 }
